+++ src/org/soulspace/overarch/adapter/render/plantuml/c4.clj
@@ -71,6 +71,13 @@
 ;;;
 ;;; Rendering
 ;;;
+(defn render-link
+  "Renders a link for the element `e`."
+  [e]
+  (when-let [link (:link e)]
+    (if (keyword? link)
+      (str ", $link=\"" (link e) "\"")
+      (str ", $link=\"" link "\""))))
 
 (defmethod puml/render-c4-element :boundary
   [model view indent e]
@@ -101,6 +108,7 @@
         (el/element-name e) "\""
         (when (:desc e) (str ", $descr=\"" (fns/to-singleline (:desc e)) "\""))
         (when (:type e) (str ", $type=\"" (:type e) "\""))
+        (render-link e)
         (when (:style e) (str ", $tags=\"" (puml/short-name (:style e)) "\""))
         ")")])
 
@@ -114,6 +122,7 @@
         (el/element-name e) "\""
         (when (:desc e) (str ", $descr=\"" (fns/to-singleline (:desc e)) "\""))
         (when (:tech e) (str ", $type=\"" (:tech e) "\""))
+        (render-link e)
         (if (:sprite e)
           (str ", $sprite=\"" (:name (puml/tech->sprite (:sprite e))) "\"")
           (when (puml/sprite? (:tech e))
@@ -131,6 +140,7 @@
         (el/element-name e) "\""
         (when (:desc e) (str ", $descr=\"" (fns/to-singleline (:desc e)) "\""))
         (when (:tech e) (str ", $techn=\"" (:tech e) "\""))
+        (render-link e)
         (if (:sprite e)
           (str ", $sprite=\"" (:name (puml/tech->sprite (:sprite e))) "\"")
           (when (puml/sprite? (:tech e))
@@ -155,6 +165,7 @@
                      (el/element-name e) "\""
                      (when (:desc e) (str ", $descr=\"" (fns/to-singleline (:desc e)) "\""))
                      (when (:tech e) (str ", $type=\"" (:tech e) "\""))
+                     (render-link e)
                      (if (:sprite e)
                        (str ", $sprite=\"" (:name (puml/tech->sprite (:sprite e))) "\"")
                        (when (puml/sprite? (:tech e))
@@ -170,6 +181,7 @@
             (el/element-name e) "\""
             (when (:desc e) (str ", $descr=\"" (:desc e) "\""))
             (when (:tech e) (str ", $type=\"" (:tech e) "\""))
+            (render-link e)
             (if (:sprite e)
               (str ", $sprite=\"" (:name (puml/tech->sprite (:sprite e))) "\"")
               (when (puml/sprite? (:tech e))
@@ -272,7 +284,6 @@
 ;;
 ;; C4 Imports
 ;;
-
 (defn render-c4-imports
   "Renders the imports for the diagram."
   [view]
@@ -285,7 +296,6 @@
 ;;;
 ;;; Diagram Styles
 ;;;
-
 (def c4-styles-hierarchy
   "Hierarchy for style methods."
   (-> (make-hierarchy)
@@ -330,7 +340,6 @@
 ;;;
 ;;; Diagram Layout
 ;;;
-
 (defn render-c4-layout
   "Renders the layout for the C4 diagram."
   [model view]
@@ -376,5 +385,3 @@
               (map #(puml/render-c4-element model view 0 %) rendered)
               (render-c4-legend view)
               "@enduml"])))
-
-