+++ src/org/soulspace/overarch/adapter/template/comb.clj
@@ -26,15 +26,18 @@
         ")?"
         "(.*)\\z")))
 
-(defn emit-string [s]
+(defn emit-string
+  [s]
   (print "(print " (pr-str s) ")"))
 
-(defn emit-expr [^String expr]
+(defn emit-expr
+  [^String expr]
   (if (.startsWith expr "=")
     (print "(print (str " (subs expr 1) "))")
     (print expr)))
 
-(defn- parse-string [src]
+(defn- parse-string
+  [src]
   (with-out-str
     (print "(do ")
     (loop [src src]
@@ -49,8 +52,6 @@
 ;;;
 ;;; Comb templates evaluated with core.eval
 ;;;
-
-; TODO use parsed source
 (defn compile-fn
   [args parsed]
   (core/eval
@@ -85,7 +86,7 @@
   ([engine-key parsed-template data]
    (eval parsed-template data)))
 
-(comment
+(comment ; Comb
   (defn greet-code [x] (print "Hello" x "!"))
 
   (core/eval (read-string "el/technical-architecture-node-types"))
@@ -133,9 +134,8 @@
 (def ctx (sci/init sci-opts))
 
 (defn eval-sci
-  "Evaluate a template using the supplied bindings. The template source may
-  be a string, or an I/O source such as a File, Reader or InputStream."
-  ([parsed data]
+  "Evaluate a `parsed-template` using the supplied `data` bindings."
+  ([parsed-template data]
    (let [e (:e data)
          ctx (:ctx data)
          model (:model data)
@@ -147,12 +147,16 @@
                                                                 'e e
                                                                 'view view}})]
      (try
-       (sci/with-out-str (sci/eval-string parsed opts))
+       (sci/with-out-str (sci/eval-string parsed-template opts))
        (catch Exception e
          (println "Exception while generating for template" (:template ctx))
          (println (ex-message e))
          (println (ex-data e))
-         (println (ex-cause e)))))))
+         (when (:debug ctx)
+           (println "Parsed Template: ")
+           (println parsed-template))
+         ;(println (ex-cause e))
+         )))))
 
 (defmethod t/parse-template :combsci
   ([engine-key template]
@@ -164,8 +168,7 @@
   ([engine-key parsed-template data]
    (eval-sci parsed-template data)))
 
-(comment
-  ;
+(comment ; Comb SCI
   (sci/eval-string* ctx "\"Hello\"")
   (sci/eval-string* ctx "\"Hello<% (dotimes [x 3] %> World<% ) %>!\"")
   (sci/eval-string (parse-string "Hello<% (dotimes [x 3] %> World<% ) %>!") (sci-opts))