+++ src/org/soulspace/overarch/domain/model.clj
@@ -255,13 +255,6 @@
         (recur (parent model p)))
       false)))
 
-; TODO remove, if other implementation works
-#_(defn descendant-nodes
-  "Returns the set of descendants of the node `e`."
-  [model e]
-  (when (el/model-node? (resolve-element model e))
-    (traverse-with-model model el/model-node? el/tree->set (:ct e))))
-
 (defn descendant-nodes
   "Returns the set of descendants of the node `e`."
   [model e]
@@ -321,6 +314,102 @@
        (into #{}
              (referred-xf model #(contains? el/architecture-dependency-relation-types (:el %))))))
 
+(defn requests-incoming
+  "Returns the requests relations issued by client `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referred-id->relations model))
+       (filter #(= :request (:el %)))))
+
+(defn requests-outgoing
+  "Returns the requests relations served by service `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referrer-id->relations model))
+       (filter #(= :request (:el %)))))
+
+(defn responses-incoming
+  "Returns the requests relations served by service `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referred-id->relations model))
+       (filter #(= :response (:el %)))))
+
+(defn responses-outgoing
+  "Returns the requests relations served by service `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referrer-id->relations model))
+       (filter #(= :response (:el %)))))
+
+(defn sends-incoming
+  "Returns the send relations of receiver `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referred-id->relations model))
+       (filter #(= :sends (:el %)))))
+
+(defn sends-outgoing
+  "Returns the send relations of sender `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referrer-id->relations model))
+       (filter #(= :sends (:el %)))))
+
+(defn publishes-incoming
+  "Returns the publish relations of `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referred-id->relations model))
+       (filter #(= :publish (:el %)))))
+
+(defn publishes-outgoing
+  "Returns the publish relations of `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referrer-id->relations model))
+       (filter #(= :publish (:el %)))))
+
+(defn subscribes-incoming
+  "Returns the subscribe relations of `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referred-id->relations model))
+       (filter #(= :subscribe (:el %)))))
+
+(defn subscribes-outgoing
+  "Returns the subscribe relations of `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referrer-id->relations model))
+       (filter #(= :subscribe (:el %)))))
+
+(defn dataflows-incoming
+  "Returns the incoming dataflow relations served by service `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referred-id->relations model))
+       (filter #(= :dataflow (:el %)))))
+
+(defn dataflows-outgoing
+  "Returns the outgoing dataflow relations of `e` in the `model`."
+  [model e]
+  (->> e
+       (:id)
+       (get (:referrer-id->relations model))
+       (filter #(= :dataflow (:el %)))))
+
 (defn subscribers-of
   "Returns the subscribers of the queue or publish relation `e` in the `model`."
   [model e]
@@ -350,7 +439,7 @@
          (into #{} (referred-xf model #(= :publish (:el %)))))))
 
 ;;
-;; class model
+;; code model
 ;;
 (defn superclasses
   "Returns the set of direct superclasses of the class element `e` in the `model`."
@@ -770,7 +859,9 @@
     (= :subtypes k)              (partial el/subtypes? v)
     (= :maturity? k)             (partial el/maturity-check? v)
     (= :maturity k)              (partial el/maturity? v)
+    (= :maturities k)            (partial el/maturities? v)
     (= :external? k)             (partial el/external-check? v)
+    (= :synthetic? k)            (partial el/synthetic-check? v)
     (= :name? k)                 (partial el/name-check? v)
     (= :name k)                  (partial el/name? v)
     (= :name-prefix k)           (partial el/name-prefix? v)