+++ test/org/soulspace/overarch/domain/views/code_view_test.clj
@@ -1,7 +1,7 @@
-(ns org.soulspace.overarch.domain.views.class-view-test
+(ns org.soulspace.overarch.domain.views.code-view-test
   (:require [clojure.test :refer :all]
             [org.soulspace.overarch.domain.view :refer :all]
-            [org.soulspace.overarch.domain.views.class-view :refer :all]
+            [org.soulspace.overarch.domain.views.code-view :refer :all]
             [org.soulspace.overarch.domain.model :as model]))
 
 (def test-input
@@ -10,7 +10,7 @@
 
 (deftest render-model-element?-test
   (testing "render-model-element? true"
-    (are [x y] (= x (boolean (render-model-element? test-model {:el :class-view} y)))
+    (are [x y] (= x (boolean (render-model-element? test-model {:el :code-view} y)))
       true {:el :package}
       true {:el :class}
       true {:el :field}
@@ -29,7 +29,7 @@
       true {:el :function}
       true {:el :protocol}))
   (testing "render-model-element? false"
-    (are [x y] (= x (boolean (render-model-element? test-model {:el :class-view} y)))
+    (are [x y] (= x (boolean (render-model-element? test-model {:el :code-view} y)))
       false {:el :state-machine}
       false {:el :start-state}
       false {:el :state}